samples/5-1_wikipedia/requirements.txt
mecab-python3
nltk
